Refactor methods in game-context-message component
In this commit, the addMessage() method in the game-context-message.component.ts has been refactored. Firstly, the property 'utterance' has switched places with 'response'. Secondly, the 'reset' method of 'messageForm' now takes in an object with 'botMessage' and 'userMessage' set as empty strings to ensure a clean slate for new messages.